window bug fixes
I switched milk to use "fake" fullscreen mode, which is basically just a borderless window. Now we don't force the os to change video modes when toggling fullscreen, which was causing odd resolution issues.

I also "fixed" the issue where throwing an error when in fullscreen mode would completely lock the game, forcing you to either shutdown or sign out lol. Now, when an error is thrown, the window is minimized. When done reading the error, the window is restored back to it's previous state.